Clean up code so the auth logic is now in its own controller instead of with the routes.
