+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules
 
+# DO NOT REMOVE ANYTHING BELOW THIS LINE
 # API Keys and Secrets
 .env
