+++ src/index.js
@@ -3,7 +3,13 @@
 var express = require('express');
 var bodyParser = require('body-parser');
 var passport = require('passport');
-require('dotenv').config(); // GET CONFIG
+
+console.log(process.env.NODE_ENV);
+
+// load local .env config if in development
+if(process.env.NODE_ENV === 'development') {
+	require('dotenv').config(); // GET CONFIG
+}
 
 var port = 3000;
 var app = express();
